requirements.txt
kivy==2.2.1
requests==2.31.0
openai>=1.12.0
httpx>=0.27.0
pypdf>=3.17.0
python-docx>=1.1.0
